Create a Deck Object!
Created a dynamic network object, which spawns a player prefab to handle all loading.

That prefab spawns in a deck based on a txt file that will contain a list of cards in it. It then generates a random card when it is clicked on. Have to add the code to remove it from the array